Add secure_buffer attribute
Secure-C's `secure_buffer` attribute can be applied to functions to
specify the length of pointer parameters. This information is then used
to check accesses to these pointers. `secure_buffer` takes two
parameters:
1. Target parameter (must be pointer type)
2. Length (must be integer type)

For example:

```c
int get5(int *_Nonnull buf)
    __attribute__((secure_buffer(buf, 10))) {
  return buf[5];
}
```

In the above example, the length of `buf` is given by `10`. When `buf`
is accessed, we must be able to statically prove that the offset is
within the length of the buffer or else an error should be printed.

Fixes #101