[securify] Handle function decls in same file
If a function is declared and the definition is in the same file, then
process the definition at the site of the declaration and annotate the
parameters for both. That way, callers that rely on the declaration will
have the proper analysis information at the time the call-site is
processed.

See #23