[securify] Add header file test
Nullability information can be passed through header files. If the
header file with a declaration for a function is included in the C file
with the definition of that function, then it will already be updated
with the nullability information. If it is not included, use the
compiler flag, `-include <header>` to include it, as shown in this test
case.

Fixes #23