Implement debug mode (`-mode=debug`)
When debug mode is enabled, instead of simply reporting errors for
unsafe non-null promotions or unsafe accesses, the tool will insert a
run-time check to fix the problem.

The run-time check is implemented in `secure_c.h` which is included in
any files that are modified. This file shoudl be installed somewhere
included in the include path.

Fixes #57