Add `secure-cc` and `securify-cc`
These tools provide drop-in replacements for the compiler, allowing
easier integration with existing build systems and IDEs. `secure-cc`
can be used to run the Secure-C checks before compiling your code, and
`securify-cc` can be used to insert annotations with `securify` and
insert run-time checks with `secure-c` before compiling the code. See
the
[wiki](http://rnd-github-us.huawei.com/se-tools/secure-c/wiki/Tools)
for more details.